update definitions for objimpl.h
