+++ newsfragments/3403.added.md
@@ -1,0 +1,1 @@
+Add FFI definitions `PyObject_GC_IsTracked` and `PyObject_GC_IsFinalized` on Python 3.9 and up (PyPy 3.10 and up).
