+++ newsfragments/3403.fixed.md
@@ -1,0 +1,1 @@
+Disable removed FFI definitions `_Py_GetAllocatedBlocks`, `_PyObject_GC_Malloc`, and `_PyObject_GC_Calloc` on Python 3.11 and up.
