+++ pyo3-ffi/src/objimpl.rs
@@ -1,7 +1,8 @@
-use crate::object::*;
-use crate::pyport::Py_ssize_t;
 use libc::size_t;
 use std::os::raw::{c_int, c_void};
+
+use crate::object::*;
+use crate::pyport::Py_ssize_t;
 
 extern "C" {
     #[cfg_attr(PyPy, link_name = "PyPyObject_Malloc")]
@@ -12,8 +13,12 @@ extern "C" {
     #[cfg_attr(PyPy, link_name = "PyPyObject_Free")]
     pub fn PyObject_Free(ptr: *mut c_void);
 
-    #[cfg(not(Py_LIMITED_API))]
-    pub fn _Py_GetAllocatedBlocks() -> Py_ssize_t;
+    // skipped PyObject_MALLOC
+    // skipped PyObject_REALLOC
+    // skipped PyObject_FREE
+    // skipped PyObject_Del
+    // skipped PyObject_DEL
+
     #[cfg_attr(PyPy, link_name = "PyPyObject_Init")]
     pub fn PyObject_Init(arg1: *mut PyObject, arg2: *mut PyTypeObject) -> *mut PyObject;
     #[cfg_attr(PyPy, link_name = "PyPyObject_InitVar")]
@@ -22,10 +27,19 @@ extern "C" {
         arg2: *mut PyTypeObject,
         arg3: Py_ssize_t,
     ) -> *mut PyVarObject;
+
+    // skipped PyObject_INIT
+    // skipped PyObject_INIT_VAR
+
     #[cfg_attr(PyPy, link_name = "_PyPyObject_New")]
     pub fn _PyObject_New(arg1: *mut PyTypeObject) -> *mut PyObject;
     #[cfg_attr(PyPy, link_name = "_PyPyObject_NewVar")]
     pub fn _PyObject_NewVar(arg1: *mut PyTypeObject, arg2: Py_ssize_t) -> *mut PyVarObject;
+
+    // skipped PyObject_New
+    // skipped PyObject_NEW
+    // skipped PyObject_NewVar
+    // skipped PyObject_NEW_VAR
 
     pub fn PyGC_Collect() -> Py_ssize_t;
 
@@ -40,56 +54,20 @@ extern "C" {
     #[cfg(Py_3_10)]
     #[cfg_attr(PyPy, link_name = "PyPyGC_IsEnabled")]
     pub fn PyGC_IsEnabled() -> c_int;
+
+    // skipped PyUnstable_GC_VisitObjects
 }
 
-#[cfg(not(any(Py_LIMITED_API, PyPy)))]
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub struct PyObjectArenaAllocator {
-    pub ctx: *mut c_void,
-    pub alloc: Option<extern "C" fn(ctx: *mut c_void, size: size_t) -> *mut c_void>,
-    pub free: Option<extern "C" fn(ctx: *mut c_void, ptr: *mut c_void, size: size_t)>,
-}
-
-#[cfg(not(any(Py_LIMITED_API, PyPy)))]
-impl Default for PyObjectArenaAllocator {
-    #[inline]
-    fn default() -> Self {
-        unsafe { std::mem::zeroed() }
-    }
-}
-
-extern "C" {
-    #[cfg(not(any(Py_LIMITED_API, PyPy)))]
-    pub fn PyObject_GetArenaAllocator(allocator: *mut PyObjectArenaAllocator);
-    #[cfg(not(any(Py_LIMITED_API, PyPy)))]
-    pub fn PyObject_SetArenaAllocator(allocator: *mut PyObjectArenaAllocator);
-}
-
-/// Test if a type has a GC head
 #[inline]
 pub unsafe fn PyType_IS_GC(t: *mut PyTypeObject) -> c_int {
     PyType_HasFeature(t, Py_TPFLAGS_HAVE_GC)
 }
 
-/// Test if an object has a GC head
-#[inline]
-#[cfg(not(Py_LIMITED_API))]
-pub unsafe fn PyObject_IS_GC(o: *mut PyObject) -> c_int {
-    (PyType_IS_GC(Py_TYPE(o)) != 0
-        && match (*Py_TYPE(o)).tp_is_gc {
-            Some(tp_is_gc) => tp_is_gc(o) != 0,
-            None => true,
-        }) as c_int
-}
-
 extern "C" {
     pub fn _PyObject_GC_Resize(arg1: *mut PyVarObject, arg2: Py_ssize_t) -> *mut PyVarObject;
 
-    #[cfg(not(Py_LIMITED_API))]
-    pub fn _PyObject_GC_Malloc(size: size_t) -> *mut PyObject;
-    #[cfg(not(Py_LIMITED_API))]
-    pub fn _PyObject_GC_Calloc(size: size_t) -> *mut PyObject;
+    // skipped PyObject_GC_Resize
+
     #[cfg_attr(PyPy, link_name = "_PyPyObject_GC_New")]
     pub fn _PyObject_GC_New(arg1: *mut PyTypeObject) -> *mut PyObject;
     #[cfg_attr(PyPy, link_name = "_PyPyObject_GC_NewVar")]
@@ -98,18 +76,16 @@ extern "C" {
     pub fn PyObject_GC_UnTrack(arg1: *mut c_void);
     #[cfg_attr(PyPy, link_name = "PyPyObject_GC_Del")]
     pub fn PyObject_GC_Del(arg1: *mut c_void);
+
+    // skipped PyObject_GC_New
+    // skipped PyObject_GC_NewVar
+
+    #[cfg(any(all(Py_3_9, not(PyPy)), Py_3_10))] // added in 3.9, or 3.10 on PyPy
+    #[cfg_attr(PyPy, link_name = "PyPyObject_GC_IsTracked")]
+    pub fn PyObject_GC_IsTracked(arg1: *mut PyObject) -> c_int;
+    #[cfg(any(all(Py_3_9, not(PyPy)), Py_3_10))] // added in 3.9, or 3.10 on PyPy
+    #[cfg_attr(PyPy, link_name = "PyPyObject_GC_IsFinalized")]
+    pub fn PyObject_GC_IsFinalized(arg1: *mut PyObject) -> c_int;
 }
 
-/// Test if a type supports weak references
-#[inline]
-#[cfg(not(Py_LIMITED_API))]
-pub unsafe fn PyType_SUPPORTS_WEAKREFS(t: *mut PyTypeObject) -> c_int {
-    ((*t).tp_weaklistoffset > 0) as c_int
-}
-
-#[inline]
-#[cfg(not(Py_LIMITED_API))]
-pub unsafe fn PyObject_GET_WEAKREFS_LISTPTR(o: *mut PyObject) -> *mut *mut PyObject {
-    let weaklistoffset = (*Py_TYPE(o)).tp_weaklistoffset;
-    o.offset(weaklistoffset) as *mut *mut PyObject
-}
+// skipped Py_VISIT
